+++ dto/src/main/kotlin/org/taktik/icure/services/external/rest/v1/dto/embed/MedicationDto.kt
@@ -21,7 +21,7 @@ data class MedicationDto(
 	val medicinalProduct: MedicinalproductDto? = null,
 	val numberOfPackages: Int? = null,
 	val batch: String? = null,
-	val expirationDate: Int? = null,
+	val expirationDate: Long? = null,
 	val instructionForPatient: String? = null,
 	val instructionForReimbursement: String? = null,
 	val commentForDelivery: String? = null,
